big refactoring. added string runtime
