+++ src/types/type.enum.ts
@@ -1,0 +1,16 @@
+export enum DataType {
+    Null,    // 0
+    Void,    // 1
+    Any,     // 2
+    Boolean, // 3
+    Number,  // 4
+    String,  // 5
+    List,    // 6
+    Dictionary, // 7
+    Object,   // 8
+    Class,    // 9
+    Function, // 10
+    Lambda,   // 11
+    Range,    // 12
+    Return    // 13
+}
